Make ./gradlew executable
It looks like the permissions for `./gradlew` are stored in git
(who knew?), so a fresh checkout of the project (on OSX, at least
has the file as non-executable).  This commit adds the
user-execute permission for the build.

Note that `./gradlew.bat` was already executable, so Windows users
should be in good shape already.