Add publically publish Oracle dependency from JCenter
I don't know how, but the Atlassean folks have managed to publish
a recent version of the ojdbc drivers under the com.oracle
ageis to Bintray via JCenter.  Since the company is as big as
Atlassean, it looks like it might be legit, which would allow
folks to build the project without access to the University of
Minnesota's Nexus installation.